+++ Assets/CYH/CYH_Animation/AnimationClips/Skill_0.anim.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 914b35ea9cdb67b4f8752c92b34dd355
+guid: 9dc24975bfac1894ba19e71ec8637fa7
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 7400000
