+++ Assets/CYH/CYH_Prefabs/Skill_1_HitBox.prefab.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 5e134a7dd935a7c418583f6c5ced9613
-NativeFormatImporter:
+guid: ebcc805d406376046bb3ef68782c91af
+PrefabImporter:
   externalObjects: {}
-  mainObjectFileID: 7400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
