+++ Assets/CYH/CYH_SO/Skill_1.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 5e134a7dd935a7c418583f6c5ced9613
+guid: 6cea0e09a7f8e484d9d88e32a9498a2d
 NativeFormatImporter:
   externalObjects: {}
-  mainObjectFileID: 7400000
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
