--- Assets/CYH/CYH_Scripts/Skill/SkillLogic_0_0.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c238eadf819680f4d84412605ca3eb66
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
