+++ Assets/CYH/CYH_Scripts/Skill/SkillLogic_0_HitBox.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c76b15a7616e0e64b9641a3d181f3055
+guid: c238eadf819680f4d84412605ca3eb66
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
