+++ Assets/CYH/CYH_Scripts/Skill/SkillLogic_1.cs
@@ -1,72 +1,123 @@
+using System.Collections;
+using System.Collections.Generic;
 using UnityEngine;
 
 public class SkillLogic_1 : MonoBehaviour
 {
-    [SerializeField] private ActiveSkillData data;
-    [SerializeField] private Animator animator;
+    [SerializeField] private ActiveSkillData _data;
+    [SerializeField] private PlayerControllerTypeA_Copy _playerController;
 
-    private CapsuleCollider2D SwordCollider;
+    [SerializeField] private CircleCollider2D _hitBox;
+
+    [SerializeField] private int _skillLevel = 0;
+
+    private Animator _animator;
+    private bool _isCooldown = false;
+
+    [SerializeField] private List<GameObject> _hitMonsters = new List<GameObject>();
 
 
     private void Awake()
     {
-        SwordCollider = GetComponent<CapsuleCollider2D>();
-        SwordCollider.enabled = false;
+        _animator = GetComponent<Animator>();
+        // 게임 스타트 -> _hitBox collider 끔
+        _hitBox.enabled = false;
     }
 
     private void Update()
     {
-        if (Input.GetKeyDown(KeyCode.Alpha3))
+        if (Input.GetKeyDown(KeyCode.Alpha2))
         {
-            UseSkill();
+            UseSkill(transform);
         }
     }
-    public void UseSkill()
+
+    public void UseSkill(Transform attacker)
     {
-        OnAttackStart();
-        Debug.Log("스킬사용");
+        // 쿨타임이면 return
+        if (_isCooldown) return;
+
+        // 쿨타임 체크 시작
+        _isCooldown = true;
+        StartCoroutine(CooldownCoroutine());
+
+        // 스킬 발동 전 몬스터 목록 초기화
+        _hitMonsters.Clear();
+
+        EnableHitbox();
         AnimationPlay();
+        Debug.Log("스킬사용");
     }
 
-    public void OnAttackStart()
+    public void UseSkill(Transform attacker, Transform defender)
     {
-        SwordCollider.enabled = true;
-        Debug.Log("콜라이더 킴");
+        // 스킬 발동 전 몬스터 목록 초기화
+        _hitMonsters.Clear();
+
+        EnableHitbox();
+        AnimationPlay();
+        Debug.Log("스킬사용 1타");
     }
 
-    public void OnAttackEnd()
+    public void EnableHitbox()
     {
-        SwordCollider.enabled = false;
+        _hitBox.enabled = true;
+        Debug.Log("히트박스 켜짐");
+    }
+
+
+    public void DisableHitbox()
+    {
+        _hitBox.enabled = false;
         Debug.Log("콜라이더 끔");
+
+        // 몬스터 TakeDamage 처리
+        Damage();
     }
 
 
     public void AnimationPlay()
     {
-        if (!SwordCollider.enabled)
+        if (!_hitBox.enabled)
             return;
         else
         {
-            animator.SetTrigger("UseSkill");
+            _animator.SetTrigger("UseSkill_1");
         }
     }
 
     private void OnTriggerEnter2D(Collider2D other)
     {
-        if (other.CompareTag("Monster"))
+
+        if (!other.CompareTag("Monster")) return;
+        if (!_hitMonsters.Contains(other.gameObject))
         {
-            Debug.Log("몬스터 맞음");
-            // 데미지 로직 호출
+            _hitMonsters.Add(other.gameObject);
+        }
+        Debug.Log($"몬스터 맞음");
+    }
+
+    private void Damage()
+    {
+        float damage = _playerController.AttackPoint * (0.75f + 0.0075f * _skillLevel);
+
+        foreach (var monster in _hitMonsters)
+        {
+            monster.GetComponent<Monster_CYH>().TakeDamage(damage);
+            Debug.Log($"{monster.name}에게 {damage}의 피해를 가했음");
         }
     }
 
-    private void OnEnable()
+    private IEnumerator CooldownCoroutine()
     {
-        SkillEvent.OnAAnimationEnd += OnAttackEnd;
-    }
-
-    private void OnDisable()
-    {
-        SkillEvent.OnAAnimationEnd -= OnAttackEnd;
+        float remaining = _data.CoolTime;
+        while (remaining > 0f)
+        {
+            Debug.Log($"쿨타임 남음: {remaining}초");
+            yield return new WaitForSeconds(1f);
+            remaining -= 1f;
+        }
+        _isCooldown = false;
+        Debug.Log("쿨타임 종료");
     }
 }
