+++ Assets/CYH/CYH_Scripts/Skill/SkillLogic_1.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7308b36ed7b9d53409e6aa57c1e185cd
+guid: 21413a6c049b7734c8fd92d9c0c21c96
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
